LRCI-3385 Change more references to integers into longs

+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/testray/BaseTestrayServer.java
@@ -79,7 +79,7 @@ public abstract class BaseTestrayServer implements TestrayServer {
 	}
 
 	@Override
-	public TestrayProject getTestrayProjectByID(int projectID) {
+	public TestrayProject getTestrayProjectByID(long projectID) {
 		_initTestrayProjects();
 
 		return _testrayProjectsByID.get(projectID);
